--- a/Kitodo/src/main/java/org/kitodo/production/services/data/ProcessService.java
+++ b/Kitodo/src/main/java/org/kitodo/production/services/data/ProcessService.java
@@ -279,25 +279,18 @@ public class ProcessService extends ProjectSearchService<Process, ProcessDTO, Pr
 
     @Override
     public void save(Process process) throws DataException {
-        WorkflowControllerService.updateProcessSortHelperStatus(process);
-        if (Objects.nonNull(process.getParent())) {
-            save(process.getParent());
-        }
-        super.save(process);
-        if (Objects.nonNull(process.getParent())) {
-            save(process.getParent());
-        }
+        this.save(process, false);
     }
 
     @Override
     public void save(Process process, boolean updateRelatedObjectsInIndex) throws DataException {
         WorkflowControllerService.updateProcessSortHelperStatus(process);
-        if (Objects.nonNull(process.getParent())) {
-            save(process.getParent(), updateRelatedObjectsInIndex);
-        }
+        
         super.save(process, updateRelatedObjectsInIndex);
-        if (Objects.nonNull(process.getParent())) {
-            save(process.getParent(), updateRelatedObjectsInIndex);
+
+        // save parents of process in order to refresh ElasticSearch index for parents
+        for (Process parent : findParentProcesses(process)) {
+            super.save(parent, updateRelatedObjectsInIndex);
         }
     }
 
@@ -308,6 +301,22 @@ public class ProcessService extends ProjectSearchService<Process, ProcessDTO, Pr
         enrichProcessData(process, false);
 
         super.saveToIndex(process, forceRefresh);
+    }
+
+    /**
+     * Find all parent processes for a process .
+     * 
+     * @param process the process whose parents are to be found
+     * @return the list of parent processes (direct parents and grand parents, and more)
+     */
+    public List<Process> findParentProcesses(Process process) {
+        List<Process> parents = new ArrayList<Process>();
+        Process current = process;
+        while(Objects.nonNull(current.getParent())) {
+            current = current.getParent();
+            parents.add(current);
+        }
+        return parents;
     }
 
     private int getNumberOfImagesForIndex(Workpiece workpiece) {