--- a/Kitodo/src/main/java/org/kitodo/production/services/workflow/WorkflowControllerService.java
+++ b/Kitodo/src/main/java/org/kitodo/production/services/workflow/WorkflowControllerService.java
@@ -431,7 +431,6 @@ public class WorkflowControllerService {
         Integer numberOfFiles = ServiceManager.getFileService().getNumberOfFiles(imagesOrigDirectory);
         if (!process.getSortHelperImages().equals(numberOfFiles)) {
             process.setSortHelperImages(numberOfFiles);
-            ServiceManager.getProcessService().save(process);
         }
 
         ServiceManager.getProcessService().save(process);
@@ -449,6 +448,7 @@ public class WorkflowControllerService {
     private void closeParent(Process process) throws DataException {
         if (Objects.nonNull(process.getParent()) && allChildrenClosed(process.getParent())) {
             process.getParent().setSortHelperStatus("100000000");
+            logger.info("calling save in closeParent for process id=" + process.getParent().getId());
             ServiceManager.getProcessService().save(process.getParent());
             closeParent(process.getParent());
         }
@@ -707,6 +707,7 @@ public class WorkflowControllerService {
         for (Process processForStatus : processes) {
             try {
                 setTasksStatusDown(processForStatus);
+                logger.info("calling save in setTaskStatusDownForProcesses for process id=" + processForStatus.getId());
                 ServiceManager.getProcessService().save(processForStatus, true);
                 updateProcessSortHelperStatus(processForStatus);
             } catch (DataException e) {