Related Issues:
- #5368

This pull request improves the performance when saving processes that have a parent process. It doesn't fix #5368 but removes a lot of unnecessary save operations previously triggered due to a double recursion in the `ProcessService.save` method.

This improves the performance for actions like "set status of process up/down". For example. for 20 selected processes (with parents) the time improves from 15.1s to 4.5s (status up) and 10.7s to 6.0s (status down).