@IkramMaalej 2022-11-03
In my local environment (OS: macOS, Browser Firefox v 106.0.2) :
- for 20 selected processes (with parents) the time improves from 43s to 19.5s when setting the processes status down  and from 46s to 11.5s when setting the processes status up.
- When creating of 54 processes (with children) the time is also reduced from 1min05s to 53s.